+++ tests/training/test_custom_modeling.py
@@ -26,6 +26,7 @@ from transformers import AutoConfig, AutoModelForCausalLM, AutoTokenizer
 
 import optimum
 import optimum.neuron.models.training
+from optimum.neuron.models.training.transformations_utils import GQAQKVColumnParallelLinearSpec
 from optimum.neuron.models.training.config import TrainingNeuronConfig
 from optimum.neuron.models.training.granite.modeling_granite import GraniteForCausalLM
 from optimum.neuron.utils.import_utils import (
@@ -336,4 +337,188 @@ def test_custom_modeling_matches_original(
         attn_implementation,
         monkeypatch,
     )
+    launch_procs(run_fn, world_size, tp_size, pp_size)
+
+@pytest.mark.parametrize(
+    "tp_size,num_attention_heads,num_key_value_heads,kv_size_multiplier,ground_truth",
+    [
+        [
+            8,
+            32,
+            4,
+            2,
+            [
+                [0, 1, 2, 3],
+                [8, 9, 10, 11],
+                [16, 17, 18, 19],
+                [24, 25, 26, 27],
+                [4, 5, 6, 7],
+                [12, 13, 14, 15],
+                [20, 21, 22, 23],
+                [28, 29, 30, 31],
+            ],
+        ],
+        [
+            8,
+            32,
+            4,
+            4,
+            [
+                [0, 1, 8, 9],
+                [16, 17, 24, 25],
+                [2, 3, 10, 11],
+                [18, 19, 26, 27],
+                [4, 5, 12, 13],
+                [20, 21, 28, 29],
+                [6, 7, 14, 15],
+                [22, 23, 30, 31],
+            ],
+        ],
+        [
+            8,
+            32,
+            4,
+            8,
+            [
+                [0, 8, 16, 24],
+                [1, 9, 17, 25],
+                [2, 10, 18, 26],
+                [3, 11, 19, 27],
+                [4, 12, 20, 28],
+                [5, 13, 21, 29],
+                [6, 14, 22, 30],
+                [7, 15, 23, 31],
+            ],
+        ],
+        [
+            32,
+            32,
+            4,
+            8,
+            [
+                [0],
+                [8],
+                [16],
+                [24],
+                [1],
+                [9],
+                [17],
+                [25],
+                [2],
+                [10],
+                [18],
+                [26],
+                [3],
+                [11],
+                [19],
+                [27],
+                [4],
+                [12],
+                [20],
+                [28],
+                [5],
+                [13],
+                [21],
+                [29],
+                [6],
+                [14],
+                [22],
+                [30],
+                [7],
+                [15],
+                [23],
+                [31],
+            ],
+        ],
+    ],
+    ids=[
+        "32-heads-4kv-heads-kv-mul-2,one kv head per rank",
+        "32-heads-4kv-heads-kv-mul-4,multiple kv heads per rank",
+        "32-heads-4kv-heads-kv-mul-8,all kv heads per rank",
+        "tp=32,32-heads-4kv-heads-kv-mul-8,one query head per rank",
+    ],
+)
+@is_trainium_test
+def test_compute_query_indices_for_rank(
+    tp_size, num_attention_heads, num_key_value_heads, kv_size_multiplier, ground_truth
+):
+    for tp_rank in range(tp_size):
+        expected = torch.tensor(ground_truth[tp_rank])
+        computed = GQAQKVColumnParallelLinearSpec.compute_query_indices_for_rank(
+            tp_size, tp_rank, num_attention_heads, num_key_value_heads, kv_size_multiplier
+        )
+        print(f"TP rank = {tp_rank}")
+        print(f"Expected {expected}")
+        print(f"Computed {computed}")
+        torch.testing.assert_close(expected, computed)
+
+
+@is_trainium_test
+def _test_resize_embedding(tie_embeddings):
+    tp_size = get_tensor_model_parallel_size()
+    tp_group = get_tensor_model_parallel_group()
+
+    static_seed_patcher = StaticSeedPatcher(42)
+
+    config = AutoConfig.from_pretrained(LLAMA_V2_MODEL_NAME)
+    config.tie_word_embeddings = tie_embeddings
+
+    with static_seed_patcher:
+        orig_model = AutoModelForCausalLM.from_pretrained(LLAMA_V2_MODEL_NAME, config=config)
+        orig_model.eval()
+        vocab_size = orig_model.config.vocab_size
+        new_vocab_size = vocab_size + tp_size
+
+    with static_seed_patcher:
+        orig_model.resize_token_embeddings(new_vocab_size)
+
+    with static_seed_patcher:
+        model = AutoModelForCausalLM.from_pretrained(LLAMA_V2_MODEL_NAME, config=config)
+        model.eval()
+
+    with static_seed_patcher:
+        model.resize_token_embeddings(new_vocab_size)
+
+    accelerator = create_accelerator(
+        tp_size,
+        1,
+        parallelize_embeddings=True,
+        sequence_parallel_enabled=True,
+    )
+    with static_seed_patcher:
+        model = accelerator.prepare_model(model)
+
+    # First we check that the embedding weights match
+    gathered = [torch.empty_like(model.model.embed_tokens.weight) for _ in range(tp_size)]
+    torch.distributed.all_gather(gathered, model.model.embed_tokens.weight, group=tp_group)
+    gathered_embedding = torch.cat(gathered, dim=0)
+    xm.mark_step()
+    torch.testing.assert_close(orig_model.model.embed_tokens.weight, gathered_embedding.to("cpu"))
+
+    # Second we check that logits match
+    tok = AutoTokenizer.from_pretrained(LLAMA_V2_MODEL_NAME)
+    tok.pad_token = tok.eos_token
+    inputs = tok("This is a test", max_length=24, padding="max_length", return_tensors="pt")
+    inputs = {k: v.to("xla") for k, v in inputs.items()}
+    orig_model = orig_model.to("xla")
+    orig_logits = orig_model(**inputs).logits
+    xm.mark_step()
+    logits = model(**inputs).logits
+    xm.mark_step()
+    gathered = [torch.empty_like(logits) for _ in range(tp_size)]
+    torch.distributed.all_gather(gathered, logits, group=tp_group)
+    gathered_logits = torch.cat(gathered, dim=2)
+    xm.mark_step()
+    torch.testing.assert_close(orig_logits, gathered_logits)
+
+
+@is_trainium_test
+@pytest.mark.parametrize(
+    "tie_embeddings",
+    [False, True],
+    ids=["embeddings_not_tied", "tied_embeddings"],
+)
+def test_resize_embedding(tie_embeddings):
+    world_size, tp_size, pp_size = (2, 2, 1)
+    run_fn = partial(_test_resize_embedding, tie_embeddings)
     launch_procs(run_fn, world_size, tp_size, pp_size)
